--- DLLFERRAMENTA/DLLFERRAMENTA.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.1145
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DLLFERRAMENTA", "DLLFERRAMENTA\DLLFERRAMENTA.csproj", "{4BA6F724-E449-4041-B745-D42EF675DB38}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4BA6F724-E449-4041-B745-D42EF675DB38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4BA6F724-E449-4041-B745-D42EF675DB38}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4BA6F724-E449-4041-B745-D42EF675DB38}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4BA6F724-E449-4041-B745-D42EF675DB38}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8E7D4DDE-D647-4683-B700-80A008B8FF21}
-	EndGlobalSection
-EndGlobal
